EMP Project/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Werkzeug==3.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
stripe==8.0.0
PyMySQL==1.1.0
pytz==2024.1
gunicorn==21.2.0
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-Limiter==3.5.0
redis==5.0.1
cryptography==42.0.2
Pillow==10.2.0
hiredis==2.3.2 